refactor: Move File, ContentFile, StreamFile to main init for easier visualization.
